move activeNode maintenance to OHazelcastPlugin
